Refactor common form input fields
